add correction of date transfer

+++ src/main/java/com/exercise/demotransfer/business/TransferOutput.java
@@ -2,11 +2,11 @@ package com.exercise.demotransfer.business;
 
 import lombok.Data;
 
-import java.util.ArrayList;
+import java.math.BigDecimal;
 
 @Data
 public class TransferOutput  extends OutputStatus{
 
     private Double tax_collected;
-    private Double cad;
+    private BigDecimal cad;
 }
